Building under Cygwin w/ gcc-3.4.4 yielded more uint32_t \!= long unsigned printf warnings
git-svn-id: https://urjtag.svn.sourceforge.net/svnroot/urjtag/trunk@1616 b68d4a1b-bc3d-0410-92ed-d4ac073336b7

+++ urjtag/src/tap/cable/cmd_xfer.c
@@ -62,7 +62,7 @@ extend_cmd_buffer (urj_tap_cable_cx_cmd_t *cmd)
     if (cmd->buf == NULL)
     {
         urj_error_set (URJ_ERROR_OUT_OF_MEMORY, "realloc(%s,%zd) fails",
-                       "cmd->buf", cmd->buf_len);
+                       "cmd->buf", (size_t) cmd->buf_len);
     }
 
     return cmd->buf ? 1 : 0;
